requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv==1.0.0
cryptography>=43.0.0
sentence-transformers>=3.0.0
torch>=2.0.0
pyyaml>=6.0
